separated spotify_auth functions from search, set up a path for retrieving specific album info and populating show.ejs page
