+++ controllers/album_controllers.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
-const searchQuery = require('../credentials/spotify_auth');
+const searchQuery = require('../credentials/spotify_search');
+const getAlbum = require('../credentials/get_album');
 
 router.get('/', async (req, res, next) => {
   try {
@@ -12,6 +13,26 @@ router.get('/', async (req, res, next) => {
 
     return res.render('albums/index', context);
 
+  } catch (error) {
+    console.log(error);
+    req.error = error;
+    return next();
+  }
+});
+
+router.get('/:id', async (req, res, next) => {
+  try {
+    console.log('====================');
+    console.log('==========', req.params, '==========');
+    console.log('====================')
+
+    const foundAlbum = await getAlbum(req.params.id);
+
+    const context = {
+      album: foundAlbum,
+    };
+
+    return res.render('albums/show', context);
   } catch (error) {
     console.log(error);
     req.error = error;
