+++ credentials/spotify_auth.js
@@ -17,19 +17,4 @@ const getToken = async () => {
   return response.data.access_token;
 };
 
-const searchQuery = 'voodoo';
-
-const searchReturn = async () => {
-  const token = await getToken();
-  const query = await axios.get(`https://api.spotify.com/v1/search?q=${searchQuery}&type=album&market=US&limit=2`, {
-    headers: {
-      'Authorization': 'Bearer ' + token
-    }
-  });
-  console.log(query.data.albums.items[0].images[0]);
-  return query.data.albums.items;
-};
-
-searchReturn();
-
-module.exports = searchReturn;
+module.exports = getToken;
