+++ views/albums/show.ejs
@@ -7,6 +7,6 @@
   <title>Document</title>
 </head>
 <body>
-  
+  <img src="<%= album.images[0].url %>" alt="<%= album.name %> image" />
 </body>
 </html>
